+++ src/Libraries/SmartStore.Core/Domain/Common/CommonSettings.cs
@@ -13,6 +13,7 @@ namespace SmartStore.Core.Domain.Common
             MaxLogAgeInDays = 7;
             MaxNumberOfScheduleHistoryEntries = 100;
             MaxQueuedMessagesAgeInDays = 14;
+            MaxGuestsRegistrationAgeInMinutes = 1440;  // 1 day (60 * 24).
         }
 		
 		public bool UseSystemEmailForContactUsForm { get; set; }
@@ -52,5 +53,10 @@ namespace SmartStore.Core.Domain.Common
         /// Gets or sets the maximum age of sent queued messages (in days).
         /// </summary>
         public int MaxQueuedMessagesAgeInDays { get; set; }
+
+        /// <summary>
+        /// Gets or sets the maximum registration age (in minutes) for automatic deletion of guests customers.
+        /// </summary>
+        public int MaxGuestsRegistrationAgeInMinutes { get; set; }
     }
 }