+++ src/Libraries/SmartStore.Services/Customers/Tasks/DeleteGuestsTask.cs
@@ -1,26 +1,32 @@
 ﻿using System;
-using System.Threading.Tasks;
+using SmartStore.Core.Domain.Common;
 using SmartStore.Services.Tasks;
 
 namespace SmartStore.Services.Customers
 {
     /// <summary>
-    /// Represents a task for deleting guest customers
+    /// Represents a task for deleting guest customers.
     /// </summary>
-    public partial class DeleteGuestsTask : AsyncTask
+    public partial class DeleteGuestsTask : ITask
     {
         private readonly ICustomerService _customerService;
+        private readonly CommonSettings _commonSettings;
 
-        public DeleteGuestsTask(ICustomerService customerService)
+        public DeleteGuestsTask(
+            ICustomerService customerService,
+            CommonSettings commonSettings)
         {
             _customerService = customerService;
+            _commonSettings = commonSettings;
         }
 
-		public override async Task ExecuteAsync(TaskExecutionContext ctx)
-		{
-			//60*24 = 1 day
-			var olderThanMinutes = 1440; // TODO: move to settings
-			await _customerService.DeleteGuestCustomersAsync(null, DateTime.UtcNow.AddMinutes(-olderThanMinutes), true);
-		}
+        public void Execute(TaskExecutionContext ctx)
+        {
+            Guard.NotNegative(_commonSettings.MaxGuestsRegistrationAgeInMinutes, nameof(_commonSettings.MaxGuestsRegistrationAgeInMinutes));
+
+            var registrationTo = DateTime.UtcNow.AddMinutes(-_commonSettings.MaxGuestsRegistrationAgeInMinutes);
+
+            _customerService.DeleteGuestCustomers(null, registrationTo, true);
+        }
 	}
 }